Shifting Twitch request basics to a common file
- Now all the twitch requests are called in RouteUtils.commonTwitchRequest
- Fixing some error handling
- Common function returns the body as parsed json, instead of the raw data, so updated all necessary usages accordingly
